% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_authenticate}
\alias{oos_authenticate}
\title{Authenticates against an Oracle Object Store}
\usage{
oos_authenticate(id, username, password)
}
\arguments{
\item{id}{identify domain to be connected to,}

\item{username}{Oracle Object Store username}

\item{password}{Oracle Object Store pawword for username}
}
\value{
returns a list with 5 elements.  This object should be passed to other functions.
auth_token which contains the authorization token needed for other commands
url - the full url command used for curl commands
user_id - the well formed user id of the form storage domain:user
identity_domain - Identity domain used to authenticate
auth_url - url in the form used for authentication
}
\description{
Authenticates against an Oracle Object Store
}
\examples{
#my_credentials <- oos_authenticate("a9999999","brian@oracle.com","1234")
}
