% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_get_file}
\alias{oos_get_file}
\title{Downloads a file as a stream from the Oracle Object store.  Used to load files into a dataframe}
\usage{
oos_get_file(credentials, container, file_name)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name where a file resides.}

\item{file_name}{The file name to download}
}
\value{
returns the contents of the file as a result of a read.csv.  ONly tested with csvs.
}
\description{
Downloads a file as a stream from the Oracle Object store.  Used to load files into a dataframe
}
\examples{
#my_data <- oos_get_file(my_credentials,"sales" ,"data.csv")
#my_data2 <- oos_get_file(my_credentials,"sales/2004" ,"data.csv")
}
