% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_ls}
\alias{oos_ls}
\title{Lists the contents of a container and all sub containers.}
\usage{
oos_ls(credentials, container)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name to show contents.}
}
\value{
reurns a dataframe which contains the container path and file name, the size in bytes of the file and when it was last modified.
}
\description{
Lists the contents of a container and all sub containers.
}
\examples{
#oos_ls(my_credentials,"sales" )
#oos_ls(my_credentials,"sales/2004" )
}
