% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_mkdir}
\alias{oos_mkdir}
\title{Creates a new container in the Oracle object Store}
\usage{
oos_mkdir(credentials, container)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name to be added.  If container is part of another container, include the full path.}
}
\value{
returns any errors.  NULL means success.
}
\description{
Creates a new container in the Oracle object Store
}
\examples{
#oos_mkdir(my_credentials,'sales' )
}
