% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_rm}
\alias{oos_rm}
\title{removes a file from the Oracle Object Store.}
\usage{
oos_rm(credentials, container, file_name)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name where the file resides.}

\item{file_name}{the name of the file to be removed}
}
\value{
returns any errors.  NULL means success.
}
\description{
removes a file from the Oracle Object Store.
}
\examples{
#oos_rm(my_credentials,"sales","sales.csv" )
}
