% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_upload_file}
\alias{oos_upload_file}
\title{Uploads a single file to the object store}
\usage{
oos_upload_file(credentials, container, directory, file_name)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name where the file should be uploaded.}

\item{directory}{Local filesystem directory where the file resides}

\item{file_name}{the name of the file to be uploaded.}
}
\value{
returns any errors.  NULL means success.
}
\description{
Uploads a single file to the object store
}
\examples{
#oos_upload_file(my_credentials,"sales","c://data/sales","sales.csv" )
}
